Refactor active varnet
