+++ fastmri/models/__init__.py
@@ -5,6 +5,7 @@ This source code is licensed under the MIT license found in the
 LICENSE file in the root directory of this source tree.
 """
 
-from .unet import Unet
+from .active_varnet import ActiveVarNet
 from .policy import StraightThroughPolicy
-from .varnet import NormUnet, SensitivityModel, VarNet, VarNetBlock, ActiveVarNet
+from .unet import Unet
+from .varnet import NormUnet, SensitivityModel, VarNet, VarNetBlock
